# Optimized findFirstNodeAfterLine