On my local machine:

Before:
```
| Method |    Mean |   Error |   StdDev | Rank |       Gen 0 |       Gen 1 |      Gen 2 | Allocated |
|------- |--------:|--------:|---------:|-----:|------------:|------------:|-----------:|----------:|
| Format | 5.329 s | 3.598 s | 0.1972 s |    1 | 533000.0000 | 178000.0000 | 29000.0000 |    2.5 GB |
```

after:
```
| Method |    Mean |   Error |   StdDev | Rank |       Gen 0 |       Gen 1 |      Gen 2 | Allocated |
|------- |--------:|--------:|---------:|-----:|------------:|------------:|-----------:|----------:|
| Format | 4.727 s | 2.141 s | 0.1174 s |    1 | 490000.0000 | 149000.0000 | 14000.0000 |   2.35 GB |
```