@nojaf 2020-06-21
Small improvement:
```
| Method |    Mean |    Error |   StdDev | Rank |       Gen 0 |       Gen 1 |      Gen 2 | Allocated |
|------- |--------:|---------:|---------:|-----:|------------:|------------:|-----------:|----------:|
| Format | 4.627 s | 0.2655 s | 0.0146 s |    1 | 486000.0000 | 147000.0000 | 11000.0000 |   2.35 GB |
```